fix(gallery-component): added animations, better filtering and library
add: added new libraries to the project; Bootstrap and ShuffleJs, they
seem to be working fine, but we have to be careful, as these can bring
some problems at deployment.

add: new scripts and css dependencies to the angular.json file.

modify: the gallery component, to be compliant with the requirements
from the client, using the Bootstrap framework and the ShuffleJs
library.

modify: the home component, as it had the gallery component, but I moved
some of the contents to the gallery component.